Remove dataset from git tracking

+++ ultralytics/data/dataset.py
@@ -1221,27 +1221,27 @@ class SiamDataset(YOLODataset):
             support_data = deepcopy(support_label)
             
             # Convert image to tensor
-            query_data["img"] = torch.from_numpy(query_data["img"].transpose(2, 0, 1))
-            support_data["img"] = torch.from_numpy(support_data["img"].transpose(2, 0, 1))
+            query_data["img"] = torch.from_numpy(query_data["img"].transpose(2, 0, 1)).float()
+            support_data["img"] = torch.from_numpy(support_data["img"].transpose(2, 0, 1)).float()
             
-            # Extract bboxes from instances and cls from label (cls is stored separately, not in instances)
+            # Extract bboxes from instances
             if "instances" in query_data:
                 instances = query_data.pop("instances")
                 nl = len(instances)
-                query_data["bboxes"] = torch.from_numpy(instances.bboxes) if nl else torch.zeros((0, 4))
-                query_data["batch_idx"] = torch.zeros(nl)
+                query_data["bboxes"] = torch.from_numpy(instances.bboxes).float() if nl else torch.zeros((0, 4), dtype=torch.float32)
+                query_data["batch_idx"] = torch.zeros(nl, dtype=torch.long)
                 # cls is stored separately in label dict, not in instances
                 if "cls" in query_data:
-                    query_data["cls"] = torch.from_numpy(query_data["cls"]) if nl else torch.zeros((0, 1))
+                    query_data["cls"] = torch.from_numpy(query_data["cls"]).float() if nl else torch.zeros((0, 1), dtype=torch.float32)
             
             if "instances" in support_data:
                 instances = support_data.pop("instances")
                 nl = len(instances)
-                support_data["bboxes"] = torch.from_numpy(instances.bboxes) if nl else torch.zeros((0, 4))
-                support_data["batch_idx"] = torch.zeros(nl)
+                support_data["bboxes"] = torch.from_numpy(instances.bboxes).float() if nl else torch.zeros((0, 4), dtype=torch.float32)
+                support_data["batch_idx"] = torch.zeros(nl, dtype=torch.long)
                 # cls is stored separately in label dict, not in instances
                 if "cls" in support_data:
-                    support_data["cls"] = torch.from_numpy(support_data["cls"]) if nl else torch.zeros((0, 1))
+                    support_data["cls"] = torch.from_numpy(support_data["cls"]).float() if nl else torch.zeros((0, 1), dtype=torch.float32)
 
         output = query_data
         output["support_img"] = support_data["img"]
